Mostly comments and resolved some implementation questions

+++ src/Property.java
@@ -4,7 +4,7 @@ public class Property {
 
     protected COLOR_SET color;
     protected int marketPrice;
-    protected int mortgageDivisor;
+    protected double mortgageDivisor;
     protected int baseRent;
     protected int baseHouseCost;
     protected int houseSellDivisor;
@@ -12,12 +12,27 @@ public class Property {
 
     protected int numHouses = 0;  // hotel = 5
 
+    /**
+     * Creates a 'blank' property.
+     * This constructor should be used for debug purposes only.
+     */
     public Property(String name, COLOR_SET color) {
         this.name = name;
         this.color = color;
     }
 
-    public Property(String name, COLOR_SET color, int marketPrice, int mortgageDivisor, int baseRent, int baseHouseCost, int houseSellDivisor, int[] rentTable) {
+    /**
+     * Constructs a Property object with given specifications.
+     * @param name Property name.
+     * @param color Property 'Set'.
+     * @param marketPrice Market (initial purchase) price, if applicable.
+     * @param mortgageDivisor Reduction of market price for mortgaging.
+     * @param baseRent Base rent amount without buildings or sets.
+     * @param baseHouseCost Base cost per house.
+     * @param houseSellDivisor Reduction of house price for selling.
+     * @param rentTable Table of rent amounts given 1-5 houses.
+     */
+    public Property(String name, COLOR_SET color, int marketPrice, double mortgageDivisor, int baseRent, int baseHouseCost, int houseSellDivisor, int[] rentTable) {
         this(name, color);
         this.marketPrice = marketPrice;
         this.mortgageDivisor = mortgageDivisor;
@@ -27,6 +42,9 @@ public class Property {
         this.rentTable = rentTable;
     }
 
+    /**
+     * @return True if a space like Jail, Taxes, Free Parking, etc. False otherwise.
+     */
     public boolean isFunctionalOnly() {
         return this.color == COLOR_SET.FUNCTION;
     }
